+++ src/collective/solr/tests/test_indexer.py
@@ -91,7 +91,8 @@ class QueueIndexerTests(TestCase):
         output = fakehttp(self.mngr.getConnection(), response)
         # indexing sends data
         self.proc.index(Foo(id='500', name='python test doc'))
-        self.assertEqual(sortFields(str(output)), getData('add_request.txt'))
+        self.assertEqual(sortFields(str(output)), getData(
+            'add_request.txt').rstrip('\n'))
 
     def testIndexAccessorRaises(self):
         response = getData('add_response.txt')
@@ -102,7 +103,8 @@ class QueueIndexerTests(TestCase):
             raise ValueError
         self.proc.index(Foo(id='500', name='python test doc',
                             text=brokenfunc))   # indexing sends data
-        self.assertEqual(sortFields(str(output)), getData('add_request.txt'))
+        self.assertEqual(sortFields(str(output)), getData(
+            'add_request.txt').rstrip('\n'))
 
     def testPartialIndexObject(self):
         foo = Foo(id='500', name='foo', price=42.0)
@@ -165,7 +167,8 @@ class QueueIndexerTests(TestCase):
         output = fakehttp(self.mngr.getConnection(), response)
         # reindexing sends data
         self.proc.reindex(Foo(id='500', name='python test doc'))
-        self.assertEqual(sortFields(str(output)), getData('add_request.txt'))
+        self.assertEqual(sortFields(str(output)), getData(
+            'add_request.txt').rstrip('\n'))
 
     def testUnindexObject(self):
         response = getData('delete_response.txt')
@@ -173,7 +176,8 @@ class QueueIndexerTests(TestCase):
         output = fakehttp(self.mngr.getConnection(), response)
         # unindexing sends data
         self.proc.unindex(Foo(id='500', name='python test doc'))
-        self.assertEqual(str(output), getData('delete_request.txt'))
+        self.assertEqual(str(output), getData(
+            'delete_request.txt').rstrip('\n'))
 
     def testCommit(self):
         response = getData('commit_response.txt')
@@ -181,7 +185,8 @@ class QueueIndexerTests(TestCase):
         output = fakehttp(self.mngr.getConnection(), response)
         # committing sends data
         self.proc.commit()
-        self.assertEqual(str(output), getData('commit_request.txt'))
+        self.assertEqual(str(output), getData(
+            'commit_request.txt').rstrip('\n'))
 
     def testNoIndexingWithoutAllRequiredFields(self):
         response = getData('dummy_response.txt')
@@ -271,7 +276,7 @@ class FakeHTTPConnectionTests(TestCase):
 
     def setUp(self):
         self.foo = Foo(id='500', name='python test doc')
-        self.schema_request = 'GET /solr/admin/file/?file=schema.xml'
+        self.schema_request = 'GET /solr/plone/admin/file/?file=schema.xml'
         config = getConfig()
         config.atomic_updates = True
 
@@ -291,7 +296,10 @@ class FakeHTTPConnectionTests(TestCase):
         mngr.closeConnection()
         self.assertEqual(len(output), 2)
         self.failUnless(output.get().startswith(self.schema_request))
-        self.assertEqual(sortFields(output.get()), getData('add_request.txt'))
+        self.assertEqual(
+            sortFields(output.get()),
+            getData('add_request.txt').rstrip('\n')
+        )
 
     def testThreeRequests(self):
         mngr = SolrConnectionManager(active=True)
@@ -305,8 +313,10 @@ class FakeHTTPConnectionTests(TestCase):
         mngr.closeConnection()
         self.assertEqual(len(output), 3)
         self.failUnless(output.get().startswith(self.schema_request))
-        self.assertEqual(sortFields(output.get()), getData('add_request.txt'))
-        self.assertEqual(output.get(), getData('delete_request.txt'))
+        self.assertEqual(sortFields(output.get()),
+                         getData('add_request.txt').rstrip('\n'))
+        self.assertEqual(output.get(), getData(
+            'delete_request.txt').rstrip('\n'))
 
     def testFourRequests(self):
         mngr = SolrConnectionManager(active=True)
@@ -321,9 +331,12 @@ class FakeHTTPConnectionTests(TestCase):
         mngr.closeConnection()
         self.assertEqual(len(output), 4)
         self.failUnless(output.get().startswith(self.schema_request))
-        self.assertEqual(sortFields(output.get()), getData('add_request.txt'))
-        self.assertEqual(output.get(), getData('delete_request.txt'))
-        self.assertEqual(output.get(), getData('commit_request.txt'))
+        self.assertEqual(sortFields(output.get()),
+                         getData('add_request.txt').rstrip('\n'))
+        self.assertEqual(output.get(), getData(
+            'delete_request.txt').rstrip('\n'))
+        self.assertEqual(output.get(), getData(
+            'commit_request.txt').rstrip('\n'))
 
     def testExtraRequest(self):
         # basically the same as `testThreeRequests`, except it
@@ -339,8 +352,10 @@ class FakeHTTPConnectionTests(TestCase):
         mngr.closeConnection()
         self.assertEqual(len(output), 3)
         self.failUnless(output.get().startswith(self.schema_request))
-        self.assertEqual(sortFields(output.get()), getData('add_request.txt'))
-        self.assertEqual(output.get(), getData('delete_request.txt'))
+        self.assertEqual(sortFields(output.get()),
+                         getData('add_request.txt').rstrip('\n'))
+        self.assertEqual(output.get(), getData(
+            'delete_request.txt').rstrip('\n'))
 
 
 class ThreadedConnectionTests(TestCase):
@@ -388,7 +403,8 @@ class ThreadedConnectionTests(TestCase):
         mngr.closeConnection()
         mngr.setHost(active=False)
         self.assertEqual(len(log), 3)
-        self.assertEqual(sortFields(log[0]), getData('add_request.txt'))
+        self.assertEqual(sortFields(log[0]), getData(
+            'add_request.txt').rstrip('\n'))
         self.failUnless(isinstance(log[1], SolrIndexProcessor))
         self.failUnless(isinstance(log[2], SolrConnection))
         self.failUnless(isinstance(proc, SolrIndexProcessor))